Generalize autocompletion classes w/ base classes
