+++ src/main/java/com/google/idea/perf/methodtracer/MethodTracerCommandPredictor.kt
@@ -1,0 +1,79 @@
+/*
+ * Copyright 2020 Google LLC
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package com.google.idea.perf.methodtracer
+
+import com.google.idea.perf.CommandPredictor
+import com.google.idea.perf.fuzzySearch
+import com.intellij.openapi.progress.ProgressManager
+
+class MethodTracerCommandPredictor: CommandPredictor {
+    @Volatile
+    private var classNames = emptyList<String>()
+
+    fun setClasses(classes: Collection<Class<*>>) {
+        classNames = classes.mapNotNull { it.canonicalName }
+    }
+
+    override fun predict(text: String, offset: Int): List<String> {
+        val tokens = text.trimStart().split(' ', '\t')
+        val normalizedText = tokens.joinToString(" ")
+        val command = parseTracerCommand(normalizedText)
+        val tokenIndex = getTokenIndex(normalizedText, offset)
+        val token = tokens.getOrElse(tokenIndex) { "" }
+
+        return when (tokenIndex) {
+            0 -> predictToken(
+                listOf("clear", "reset", "trace", "untrace"), token
+            )
+            1 -> when (command) {
+                is TracerCommand.Trace -> {
+                    val options = predictToken(listOf("all", "count", "wall-time"), token)
+                    val classes = predictToken(classNames, token)
+                    return options + classes
+                }
+                else -> emptyList()
+            }
+            2 -> when {
+                command is TracerCommand.Trace && command.target is TraceTarget.Method ->
+                    predictMethodToken(command.target.className, token)
+                command is TracerCommand.Trace -> predictToken(classNames, token)
+                else -> emptyList()
+            }
+            3 -> when {
+                command is TracerCommand.Trace && command.target is TraceTarget.Method ->
+                    predictMethodToken(command.target.className, token)
+                else -> emptyList()
+            }
+            else -> emptyList()
+        }
+    }
+
+    private fun predictToken(choices: Collection<String>, token: String): List<String> {
+        return fuzzySearch(choices, token, -1) { ProgressManager.checkCanceled() }
+            .map { it.source }
+    }
+
+    private fun predictMethodToken(className: String, token: String): List<String> {
+        val clazz = Class.forName(className)
+        val methodNames = clazz.methods.map { it.name.substringAfter('$') }
+        return predictToken(methodNames, token)
+    }
+
+    private fun getTokenIndex(input: String, index: Int): Int {
+        return input.subSequence(0, index).count { it.isWhitespace() || it == '#' }
+    }
+}
